+++ packages/abjad-convert/src/converterFactory.ts
@@ -1,5 +1,6 @@
 import { Abjad } from './types'
 import { IConverter } from './IConverter'
+import { ArabicToHanifiRohingyaConverter } from './scripts/arabic/toHanifiRohingys'
 import { ArabicToImperialAramaicConverter } from './scripts/arabic/toImperialAramaic'
 import { ArabicToIpaConverter } from './scripts/arabic/toIpa'
 import { ArabicToOldSouthArabianConverter } from './scripts/arabic/toOldSouthArabian'
@@ -7,6 +8,7 @@ import { ArabicToPhoenicianConverter } from './scripts/arabic/toPhoenician'
 import { ArabicToSyriacConverter } from './scripts/arabic/toSyriac'
 import { ArabicToTifinaghConverter } from './scripts/arabic/toTifinagh'
 import { ArabicToUgariticConverter } from './scripts/arabic/toUgaritic'
+import { HanifiRohingyaToArabicConverter } from './scripts/hanifiRohingya/toArabic'
 import { ImperialAramaicToArabicConverter } from './scripts/imperialAramaic/toArabic'
 import { IpaToArabicConverter } from './scripts/ipa/toArabic'
 import { OldSouthArabicToArabicConverter } from './scripts/oldSouthArabian/toArabic'
@@ -16,6 +18,7 @@ import { TifinaghToArabicConverter } from './scripts/tifinagh/toArabic'
 import { UgariticToArabicConverter } from './scripts/ugaritic/toArabic'
 
 const converters: IConverter[] = [
+	new ArabicToHanifiRohingyaConverter(),
 	new ArabicToImperialAramaicConverter(),
 	new ArabicToIpaConverter(),
 	new ArabicToOldSouthArabianConverter(),
@@ -23,6 +26,7 @@ const converters: IConverter[] = [
 	new ArabicToSyriacConverter(),
 	new ArabicToTifinaghConverter(),
 	new ArabicToUgariticConverter(),
+	new HanifiRohingyaToArabicConverter(),
 	new ImperialAramaicToArabicConverter(),
 	new IpaToArabicConverter(),
 	new OldSouthArabicToArabicConverter(),