+++ packages/abjad-convert/src/types.ts
@@ -1,5 +1,6 @@
 export enum Abjad {
 	Arabic = 'arabic',
+	HanifiRohingya = 'hanifi-rohingya',
 	ImperialAramaic = 'imperial-aramaic',
 	IPA = 'ipa',
 	OldSouthArabian = 'old-south-arabian',