+++ packages/abjad-convert/CHANGELOG.md
@@ -2,6 +2,10 @@
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
 ## [0.9.0] 2024-05-15
+### Added
+- Support Hanifi Rohingya
+- letter ڠ
+- mapping to IPA for letter ژ
 
 ## [0.6.1] 2024-04-23
 ### Added